Rename otlp -> otel
