+++ .golangci.yaml
@@ -70,9 +70,9 @@ linters-settings:
         alias: kitmetricpipeline
       - pkg: github.com/kyma-project/telemetry-manager/test/testkit/kyma/telemetry/trace
         alias: kittracepipeline
-      - pkg: github.com/kyma-project/telemetry-manager/test/testkit/otlp/metrics
+      - pkg: github.com/kyma-project/telemetry-manager/test/testkit/otel/metrics
         alias: kitmetrics
-      - pkg: github.com/kyma-project/telemetry-manager/test/testkit/otlp/traces
+      - pkg: github.com/kyma-project/telemetry-manager/test/testkit/otel/traces
         alias: kittraces
       - pkg: github.com/kyma-project/telemetry-manager/webhook/logparser
         alias: logparserwebhook
@@ -203,7 +203,7 @@ issues:
     - linters: [dupl]
       path: test/testkit/matchers/(log|metric|trace)_matchers.go
     - linters: [errcheck]
-      path: test/testkit/otlp/traces/traces.go
+      path: test/testkit/otel/traces/traces.go
     - linters: [dupl]
       path: test/testkit/verifiers/(deployment|daemonset|metrics|traces).go
     - linters: [dupl]
