+++ test/testkit/otel/traces/exporter.go
